refactor(imports): standardize import paths for work-category types
