+++ app/users/dashboard/portfolio/projects/all-projects/[id]/tasks/new/page.tsx
@@ -40,7 +40,7 @@ import {
   useCreateWorkCategory,
 } from '@/hooks/work-category';
 import { useCurrentUserEmployee } from '@/hooks/employee';
-import { abbreviatedName } from '@/types/task/work-category';
+import { abbreviatedName } from '@/types/work-category';
 
 import { TaskStatus, getTaskStatusLabel } from '@/types/task/task-status';
 import { toast } from '@/lib/styles/toast-styles';
